added layouts to all pages
