+++ web/.prettierrc
@@ -1,7 +1,8 @@
 {
-  "bracketSpacing": false,
+  "bracketSpacing": true,
   "jsxBracketSameLine": true,
   "singleQuote": true,
   "trailingComma": "all",
-  "arrowParens": "avoid"
+  "arrowParens": "avoid",
+  "jsxSingleQuote": true
 }
