+++ web/Layouts/Layout.jsx
@@ -1,0 +1,10 @@
+import Header from '../components/Header';
+
+export default ({children}) => {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  );
+};
