+++ web/pages/_app.js
@@ -1,18 +1,23 @@
 import '../styles/globals.css';
 import '@elastic/eui/dist/eui_theme_light.css';
 
-import {Provider} from 'react-redux';
-import {ThemeProvider} from 'next-themes';
+import React from 'react';
+import { Provider } from 'react-redux';
+import { ThemeProvider } from 'next-themes';
 
 import store from '../store';
 import PageProvider from '../components/PageProvider';
 
-function MyApp({Component, pageProps}) {
+function MyApp({ Component, pageProps }) {
+  const Layout = Component.Layout || React.Fragment;
+
   return (
     <Provider store={store}>
       <ThemeProvider>
         <PageProvider>
-          <Component {...pageProps} />
+          <Layout>
+            <Component {...pageProps} />
+          </Layout>
         </PageProvider>
       </ThemeProvider>
     </Provider>
